requirements_txt.txt
chatterbox-tts==0.1.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
torch==2.5.1
torchaudio==2.5.1
librosa==0.10.2
huggingface_hub==0.26.5
numpy==1.26.4
scipy==1.14.1
soundfile==0.12.1
pydantic==2.10.3
python-multipart==0.0.19
aiofiles==24.1.0